+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
-
-* TODO: Describe initial release.
+## 2.0.0
+**BREAKING**: Removed the ``duration`` parameter
+- There is no need to pass ``duration`` anymore. Instead, there is a new property called ``recommendedDuration`` which is calculated based on the ``bounce`` parameter.
+- Completely changed the internal formulas used for the spring simulation.
+## 1.0.1
+- First stable release